miscellaneous details

+++ src/main/java/Observer.java
@@ -1,4 +1,4 @@
 public interface Observer {
     WeatherStation weatherStation = new WeatherStation();
-    void update(int temperature);
+    void update(double temperature);
 }
